requirements.txt
Flask==3.0.3
dependency-injector==4.41.0
pytest-cov==5.0.0
pytest-flask==1.3.0
